DMD-409 util functions
