+++ utils/file-reader.ts
@@ -1,4 +1,4 @@
-export function readFile (file: File, cb): void {
+export function readFile (file: File, cb: (e: ProgressEvent<EventTarget>) => void): void {
   const reader = new FileReader()
 
   reader.onload = (e: ProgressEvent) => {
